Move composeEnhancers to util
