+++ packages/venia-concept/src/store.js
@@ -2,7 +2,6 @@ import { createStore } from 'redux';
 
 import enhancer from 'src/middleware';
 import reducer from 'src/reducers';
-
-import composeEnhancers from './composeEnhancers';
+import composeEnhancers from 'src/util/composeEnhancers';
 
 export default createStore(reducer, composeEnhancers(enhancer));
